% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddisttest.R
\name{rddisttest}
\alias{rddisttest}
\title{Tests for Manipulation of Discrete Running Variable}
\usage{
rddisttest(RV, C, K = 0)
}
\arguments{
\item{RV}{Vector containing the values for each observations level of the running variable.}

\item{C}{Real number specifying the level of the RV at which the discontinuity exists.}

\item{K}{Real number corresponding to the user specified bound on the RV’s probability mass function curvature.}
}
\value{
P-value
}
\description{
Provides an alternative to conventional RD manipulation tests, such as the McCrary test, to be applied in the case where the running variable takes on discrete values.
}
\examples{
\dontrun{
#rddisttest(dataframe$runningvar, 0, .01)
}
}
